Styling changes, src from scratch
