+++ .eslintrc.js
@@ -1,59 +1,73 @@
 module.exports = {
-    parser: '@typescript-eslint/parser',
-    plugins: ['@typescript-eslint'],
-    extends: [
-        "eslint:recommended",
-        "plugin:@typescript-eslint/recommended",
+  parser: "@typescript-eslint/parser",
+  plugins: ["@typescript-eslint"],
+  extends: ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
+  rules: {
+    "array-bracket-spacing": ["error", "never"],
+    "block-spacing": ["error", "always"],
+    "comma-spacing": [
+      "error",
+      {
+        before: false,
+        after: true,
+      },
     ],
-	rules: {
-        'array-bracket-spacing': ['error', 'never'],
-        'block-spacing': ['error', 'always'],
-        'comma-spacing': ['error', {
-            before: false,
-            after: true,
-        }],
-        'indent': ['error', 4, {
-            'SwitchCase': 1,
-        }],
-        'jsx-quotes': ['error', 'prefer-double'],
-        'keyword-spacing': ['error', {
-            after: true,
-            before: true,
-        }],
-        'object-curly-spacing': ['error', 'never'],
-        'no-multi-spaces': 'error',
-        'no-trailing-spaces': 'error',
-        'no-whitespace-before-property': 'error',
-        'semi': ['error', 'always'],
-        'space-before-function-paren': ['error', {
-            anonymous: 'always',
-            asyncArrow: 'always',
-            named: 'never',
-        }],
-        'space-in-parens': ['error', 'never'],
-        'space-infix-ops': 'error',
-        'spaced-comment': ['error', 'always', {exceptions: ['-']}],
-        'quotes': ['error', 'single', {
-            allowTemplateLiterals: true,
-            avoidEscape: true,
-        }],
-        'eol-last': ["error", "always"],
-        '@typescript-eslint/camelcase': 'off',
-        '@typescript-eslint/explicit-function-return-type': 'off',
-        '@typescript-eslint/explicit-member-accessibility': 'off',
-        '@typescript-eslint/func-call-spacing': ['error', 'never'],
-        '@typescript-eslint/no-explicit-any': 'off',
-        '@typescript-eslint/no-object-literal-type-assertion': 'off',
-        '@typescript-eslint/no-use-before-define': 'off',
-        '@typescript-eslint/explicit-module-boundary-types': 'off',
+    indent: [
+      "error",
+      4,
+      {
+        SwitchCase: 1,
+      },
+    ],
+    "jsx-quotes": ["error", "prefer-double"],
+    "keyword-spacing": [
+      "error",
+      {
+        after: true,
+        before: true,
+      },
+    ],
+    "object-curly-spacing": ["error", "never"],
+    "no-multi-spaces": "error",
+    "no-trailing-spaces": "error",
+    "no-whitespace-before-property": "error",
+    semi: ["error", "always"],
+    "space-before-function-paren": [
+      "error",
+      {
+        anonymous: "always",
+        asyncArrow: "always",
+        named: "never",
+      },
+    ],
+    "space-in-parens": ["error", "never"],
+    "space-infix-ops": "error",
+    "spaced-comment": ["error", "always", { exceptions: ["-"] }],
+    quotes: [
+      "error",
+      "single",
+      {
+        allowTemplateLiterals: true,
+        avoidEscape: true,
+      },
+    ],
+    "eol-last": ["error", "always"],
+    "@typescript-eslint/camelcase": "off",
+    "@typescript-eslint/explicit-function-return-type": "off",
+    "@typescript-eslint/explicit-member-accessibility": "off",
+    "@typescript-eslint/func-call-spacing": ["error", "never"],
+    "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/no-object-literal-type-assertion": "off",
+    "@typescript-eslint/no-use-before-define": "off",
+    "@typescript-eslint/explicit-module-boundary-types": "off",
+  },
+  overrides: [
+    {
+      files: ["tasks/**/*.js"],
+      rules: {
+        "@typescript-eslint/no-var-requires": "off",
+        "no-undef": "off",
+      },
     },
-    overrides: [
-        {
-            files: ['tasks/**/*.js'],
-            rules: {
-                '@typescript-eslint/no-var-requires': 'off',
-                'no-undef': 'off',
-            },
-        },
-    ],
+  ],
 };
