+++ assets/manifest.json
@@ -1,10 +1,9 @@
 {
   "name": "Deadname Remover",
-
-  "version": "1.2.1",
+  "version": "2.0.0",
   "manifest_version": 2,
   "description": "An easy to use Google Chrome plugin to automatically remove and replace deadnames",
-  "homepage_url": "https://github.com/WillHayCode/Deadname-Remover",
+  "homepage_url": "https://github.com/arimgibson/Deadname-Remover",
   "icons": {
     "16": "icons/icon16.png",
     "48": "icons/icon48.png",
