--- src/index.ts
@@ -1,13 +1,0 @@
-import {start} from './inject/inject';
-
-if (!window.chrome) {
-    window.chrome = {} as any;
-}
-if (!chrome.runtime) {
-    chrome.runtime = {} as any;
-}
-
-const port = chrome.runtime.connect({name: 'tab'});
-port.onMessage.addListener((settings) => {
-    start(settings);
-});
